Generate real static pages for each episode
This is done via Jekyll's post system, and a custom rake task that
parses the remote RSS feed and creates post files for each episode.

This is better than simply copying the RSS feed into the site and having
client-side javascript actually render out the pages, as search engines
don't always like it.

That said, I will be adding Backbone.js back into the mix to allow it to
nagivate between episodes and the front-page dynamically without having
to reload the page. But at the same time directly loading the page for
an episode will work too.